requirements.txt
# Flask Core
Flask==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
Flask-Mail==0.9.1
Werkzeug==3.1.3

# Authentication & Security
bcrypt==4.2.1
email-validator==2.1.0

# MongoDB Dependencies
pymongo[srv]==4.13.2
mongoengine==0.29.1
Flask-MongoEngine==1.0.0
dnspython==2.7.0

# Utilities
python-dotenv==1.0.1
click==8.1.8
blinker==1.9.0
itsdangerous==2.2.0
MarkupSafe==3.0.2
Jinja2==3.1.5

# Production Server
gunicorn==23.0.0

# Development (optional)
# flask-cors==4.0.0  # Uncomment if using React frontend
